move input out of folder
