+++ apps/api/v2/src/ee/event-types/inputs/update-event-type.input.ts
